requirements.txt
numpy~=1.24.4
opencv-python~=4.6.0.66
boto3~=1.28.53
pillow~=10.0.1
pytesseract~=0.3.10
tqdm~=4.65.0
amazon-textract-textractor
google-cloud-vision
paddleocr~=2.7.0.3